JL-4: Added concept for new auth flow

+++ src/app/api/auth/logout/route.js
@@ -1,15 +1,16 @@
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
-import { sessionsRepo } from "~/infra/repos/sessions-repo";
+import { sessionsRepo } from "~/infra/repositories/public/sessions";
 import { PUBLIC_ROUTES } from "~/app/routes";
 
 export async function GET() {
   var cookieStore = await cookies();
   var token = cookieStore.get(process.env.COOKIE_NAME)?.value;
 
-  await sessionsRepo.destroy(token);
-
-  cookieStore.delete(process.env.COOKIE_NAME);
+  if (token) {
+    await sessionsRepo.destroy(token);
+    cookieStore.delete(process.env.COOKIE_NAME);
+  }
 
   redirect(PUBLIC_ROUTES.auth.login());
 }
